+++ keybindings.h
@@ -1,15 +1,15 @@
 struct KeyBinding {
   char * name;
-  int key; };
+  uint16_t key; };
 
 struct KeysWinData {
-  int max;
+  uint16_t max;
   char edit;
-  int select; };
+  uint16_t select; };
 
 void init_keybindings(struct World *);
 void save_keybindings(struct World *);
-int get_action_key (struct KeyBinding *, char *);
-char * get_keyname(int);
+uint16_t get_action_key (struct KeyBinding *, char *);
+char * get_keyname(uint16_t);
 void keyswin_mod_key (struct World *, struct WinMeta *);
 void keyswin_move_selection (struct World *, char);
